+++ x/domain/keeper/fees.go
@@ -1,10 +1,11 @@
 package keeper
 
 import (
+	"log"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/distribution"
 	"github.com/iov-one/iovns/x/domain/types"
-	"log"
 )
 
 // CollectFees collects the fees of a msg and sends them
@@ -23,8 +24,6 @@ func (k Keeper) CollectFees(ctx sdk.Context, msg sdk.Msg, addr sdk.AccAddress) e
 	case *types.MsgDeleteAccount:
 		level = len(msg.Domain)
 	case *types.MsgDeleteDomain:
-		level = len(msg.Domain)
-	case *types.MsgFlushDomain:
 		level = len(msg.Domain)
 	case *types.MsgRegisterAccount:
 		level = len(msg.Domain)